requirements.txt
aiogram
asyncpg
SQLAlchemy
python-dotenv
openai
Pillow